pyproject.toml
[project]
name = "bot-consulta-publica"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-telegram-bot (>=22.4,<23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
